Ensure we do not expose any cultures we do not accept in languages. Futhermore, I moved it to a service to people can replace the logic
