+++ src/Umbraco.Core/Services/ICultureService.cs
@@ -1,0 +1,8 @@
+using System.Globalization;
+
+namespace Umbraco.Cms.Core.Services;
+
+public interface ICultureService
+{
+    CultureInfo[] GetValidCultureInfos();
+}
